pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "queue-worker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "PyYAML>=6.0",
    "fastapi[standard]>=0.115",
    "playwright>=1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
queue-worker = "queue_worker.cli:main"
queue-worker-web = "queue_worker.web:main"

[tool.setuptools.packages.find]
where = ["src"]
